bug-hunt #554: NUMA balancing - fix migration to prevent hot-move bouncing tasks
Key changes:
1. Fix numa_migrate_page() to use pmm_alloc_frame_on_node() for
   node-local allocation instead of global pmm_alloc_frame()
2. Change return type from int to uint64_t — now returns the new
   physical frame address on success (caller can update page table
   and free old frame), 0 on failure
3. Add migration cool-down tracker to prevent bouncing: 32-entry
   array tracking recently-migrated physical addresses; pages on
   cool-down (4 scan cycles) are skipped during scanning, preventing
   repeated back-and-forth migration for tasks that access memory
   across multiple NUMA nodes
4. Connect numa_scan_process() to actually detect remote pages via
   phys_to_node_id() helper and migrate them to the home node, with
   cooldown check, PTE update, TLB flush (invlpg), and old-frame free
5. Update hugepage_migration.c caller to use new uint64_t return type
6. Add #include "numa_mem.h" for node-local allocator access

+++ src/include/numa_balancing.h
@@ -35,11 +35,20 @@ void numa_balancing_init(void);
 void numa_hint_fault(uint64_t addr, int node);
 
 /* Migrate a page to the target NUMA node using the PMM migration path.
- * @page:   Physical address of the page (frame number << 12)
+ * Allocates a new frame on the target node via pmm_alloc_frame_on_node(),
+ * copies the page contents, and records the migration in the cool-down
+ * tracker to prevent bouncing.
+ *
+ * @page:        Physical address of the source page
  * @target_node: Destination NUMA node ID
- * Returns 0 on success, negative errno on failure.
+ *
+ * Returns the physical address of the new frame on success,
+ * 0 on failure (allocation failure, invalid params, or !initialised).
+ *
+ * NOTE: The caller is responsible for updating the page table entry
+ * to point to the returned frame and for freeing the old frame.
  */
-int numa_migrate_page(uint64_t page, int target_node);
+uint64_t numa_migrate_page(uint64_t page, int target_node);
 
 /* Get per-node statistics.
  * @node:  NUMA node ID
